move protocolsforobject from the factory
