+++ pkg/kubectl/cmd/expose.go
@@ -37,6 +37,7 @@ import (
 	"k8s.io/kubernetes/pkg/kubectl/genericclioptions"
 	"k8s.io/kubernetes/pkg/kubectl/genericclioptions/printers"
 	"k8s.io/kubernetes/pkg/kubectl/genericclioptions/resource"
+	"k8s.io/kubernetes/pkg/kubectl/polymorphichelpers"
 	"k8s.io/kubernetes/pkg/kubectl/scheme"
 	"k8s.io/kubernetes/pkg/kubectl/util/i18n"
 )
@@ -95,7 +96,7 @@ type ExposeServiceOptions struct {
 	ClientForMapping          func(*meta.RESTMapping) (resource.RESTClient, error)
 	MapBasedSelectorForObject func(runtime.Object) (string, error)
 	PortsForObject            func(runtime.Object) ([]string, error)
-	ProtocolsForObject        func(runtime.Object) (map[string]string, error)
+	ProtocolsForObject        polymorphichelpers.ProtocolsForObjectFunc
 
 	Namespace string
 	Mapper    meta.RESTMapper
@@ -194,7 +195,7 @@ func (o *ExposeServiceOptions) Complete(f cmdutil.Factory, cmd *cobra.Command) e
 	o.ClientForMapping = f.ClientForMapping
 	o.MapBasedSelectorForObject = f.MapBasedSelectorForObject
 	o.PortsForObject = f.PortsForObject
-	o.ProtocolsForObject = f.ProtocolsForObject
+	o.ProtocolsForObject = polymorphichelpers.ProtocolsForObjectFn
 	o.Mapper, err = f.ToRESTMapper()
 	if err != nil {
 		return err
