+++ pkg/kubectl/cmd/util/BUILD
@@ -70,7 +70,6 @@ go_test(
         "//pkg/api/testapi:go_default_library",
         "//pkg/api/testing:go_default_library",
         "//pkg/apis/core:go_default_library",
-        "//pkg/kubectl:go_default_library",
         "//pkg/kubectl/genericclioptions:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/equality:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/errors:go_default_library",
